Начата разработка канбан доски

+++ src/main/java/ru/manager/ProgectManager/repositories/StatisticsRepository.java
@@ -3,6 +3,6 @@ package ru.manager.ProgectManager.repositories;
 import org.springframework.data.repository.CrudRepository;
 import ru.manager.ProgectManager.entitys.StatisticsUsing;
 
-public interface StatisticsRepository extends CrudRepository<StatisticsUsing, Integer> {
+public interface StatisticsRepository extends CrudRepository<StatisticsUsing, Long> {
     StatisticsUsing findByType(String type);
 }
