+++ src/main/java/ru/manager/ProgectManager/repositories/UserWithRoleConnectorRepository.java
@@ -3,5 +3,5 @@ package ru.manager.ProgectManager.repositories;
 import org.springframework.data.repository.CrudRepository;
 import ru.manager.ProgectManager.entitys.UserWithRoleConnector;
 
-public interface UserWithRoleConnectorRepository extends CrudRepository<UserWithRoleConnector, Integer> {
+public interface UserWithRoleConnectorRepository extends CrudRepository<UserWithRoleConnector, Long> {
 }
